# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para gerenciamento de envios de pacotes por diferentes transportadoras",
        "title": "Delivery Manager API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Verifica se a API está funcionando corretamente",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check da API",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package": {
            "get": {
                "description": "Retorna todos os pacotes cadastrados no sistema. Endpoint temporário para desenvolvimento.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Listar todos os pacotes",
                "responses": {
                    "200": {
                        "description": "Lista de todos os pacotes",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.PackageResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package/": {
            "post": {
                "description": "Cria um novo pacote com produto, peso e estado de destino. O sistema automaticamente mapeia o estado para a região correspondente e calcula as transportadoras disponíveis.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Criar um novo pacote",
                "parameters": [
                    {
                        "description": "Dados do pacote",
                        "name": "package",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.PackageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Pacote criado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "string"
                                            },
                                            "message": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou estado não suportado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package/hire-carrier": {
            "post": {
                "description": "Contrata uma transportadora para realizar a entrega do pacote. O status do pacote será automaticamente alterado para 'esperando_coleta'. Transportadoras disponíveis: nebulix, rotafacil, moventra.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Contratar transportadora",
                "parameters": [
                    {
                        "description": "Dados para contratação",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.HireCarrierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transportadora contratada com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou transportadora não disponível",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Pacote não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Pacote já possui transportadora contratada",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package/status": {
            "put": {
                "description": "Atualiza o status de um pacote específico. Status válidos: criado, esperando_coleta, coletado, enviado, entregue, extraviado.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Atualizar status de um pacote",
                "parameters": [
                    {
                        "description": "Dados para atualização de status",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.UpdateStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status atualizado com sucesso",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "allOf": [
                                    {
                                        "type": "string"
                                    },
                                    {
                                        "type": "object",
                                        "properties": {
                                            "message": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                ]
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou status não permitido",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Pacote não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package/{id}": {
            "get": {
                "description": "Retorna os dados completos de um pacote pelo ID, incluindo informações de entrega se uma transportadora foi contratada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Consultar um pacote específico",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID único do pacote",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do pacote",
                        "schema": {
                            "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.PackageResponse"
                        }
                    },
                    "404": {
                        "description": "Pacote não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/package/{id}/quote": {
            "post": {
                "description": "Retorna cotações de frete disponíveis para um pacote, ordenadas por prazo de entrega. Inclui preços e prazos estimados de todas as transportadoras que atendem a região do pacote.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "packages"
                ],
                "summary": "Cotação de fretes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID do pacote",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cotações de frete disponíveis",
                        "schema": {
                            "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.ShippingsQuoteResponse"
                        }
                    },
                    "404": {
                        "description": "Pacote não encontrado",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.HireCarrierRequest": {
            "description": "Dados necessários para contratar uma transportadora",
            "type": "object",
            "required": [
                "carrier_id",
                "package_id"
            ],
            "properties": {
                "carrier_id": {
                    "type": "string",
                    "example": "nebulix"
                },
                "package_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                }
            }
        },
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.PackageRequest": {
            "description": "Dados necessários para criar um novo pacote",
            "type": "object",
            "required": [
                "estado_destino",
                "peso_kg",
                "produto"
            ],
            "properties": {
                "estado_destino": {
                    "type": "string",
                    "example": "PR"
                },
                "peso_kg": {
                    "type": "number",
                    "maximum": 1000,
                    "example": 0.6
                },
                "produto": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 2,
                    "example": "Camisa tamanho G"
                }
            }
        },
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.PackageResponse": {
            "description": "Resposta com os dados de um pacote",
            "type": "object",
            "properties": {
                "entrega": {
                    "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.ShippingQuoteResponse"
                },
                "estado_destino": {
                    "type": "string",
                    "example": "PR"
                },
                "id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "peso_kg": {
                    "type": "number",
                    "example": 0.6
                },
                "produto": {
                    "type": "string",
                    "example": "Camisa tamanho G"
                },
                "regiao_destino": {
                    "type": "string",
                    "example": "sul"
                },
                "status": {
                    "type": "string",
                    "example": "criado"
                }
            }
        },
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.ShippingQuoteResponse": {
            "description": "Dados de uma cotação de frete",
            "type": "object",
            "properties": {
                "nome": {
                    "type": "string",
                    "example": "Nebulix Logística"
                },
                "prazo_estimado": {
                    "type": "integer",
                    "example": 4
                },
                "preco": {
                    "type": "number",
                    "example": 42.5
                },
                "transportadora_id": {
                    "type": "string",
                    "example": "nebulix"
                }
            }
        },
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.ShippingsQuoteResponse": {
            "description": "Resposta com as cotações de frete disponíveis",
            "type": "object",
            "properties": {
                "transportadoras": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.ShippingQuoteResponse"
                    }
                }
            }
        },
        "github_com_foliveiracamara_delivery-manager-api_internal_api_http_dto.UpdateStatusRequest": {
            "description": "Dados necessários para atualizar o status de um pacote",
            "type": "object",
            "required": [
                "package_id",
                "status"
            ],
            "properties": {
                "package_id": {
                    "type": "string",
                    "example": "123e4567-e89b-12d3-a456-426614174000"
                },
                "status": {
                    "type": "string",
                    "example": "enviado"
                }
            }
        }
    }
}